pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "neural-search"
version = "0.1.0"
description = "Production-grade semantic search API with embeddings and vector search"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sentence-transformers>=2.2.2",
    "torch>=2.1.0",
    "qdrant-client>=1.7.0",
    "faiss-cpu>=1.7.4",
    "rank-bm25>=0.2.2",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "prometheus-client>=0.19.0",
    "httpx>=0.26.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/neural_search"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 100
